fixed layout for schedule
